src/speed_layer/requirements.txt
snowflake-connector-python==3.8.1
